cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intelligencre-cli"
version = "1.0.0"
description = "Intelligencre Studio CLI - Databricks API from the terminal"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dbx-cli = "cli.main:cli"

[tool.setuptools.packages.find]
include = ["cli*"]
